requirements.txt
django==5.2.4
django-ninja==1.4.3
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
python-dotenv==1.1.1